refactor: CardNumber 경로 변경
